+++ src/platform_impl/windows/mod.rs
@@ -69,6 +69,7 @@ impl WindowId {
 
 #[macro_use]
 mod util;
+mod dark_mode;
 mod dpi;
 mod drop_handler;
 mod event;